+++ audio/wave.h
@@ -1,0 +1,7 @@
+#ifndef AUDIO_WAVE_H
+#define AUDIO_WAVE_H
+#include "../ws2812svr.h"
+
+void audio_wave(thread_context* context, char* args);
+
+#endif